comment: 다음 할 일 정리 및 오류 발견하여 인지. 다음 커밋부터 마이페이지 api작성 및 버그 수정 예정

+++ src/main/java/org/example/playground/domain/user/dto/UserSummaryDTO.java
@@ -9,7 +9,8 @@ import lombok.Getter;
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
 @Builder(access = AccessLevel.PRIVATE)
 //유저 요약 보기 (게시글, 댓글 에 나오는 정보)
+//관리자의 유저 모아보기
 public class UserSummaryDTO {
-    private Integer id;
+    private Long id;
     private String name;
 }
